Initialization flow and position and data type from int to Int64

+++ BankAccount.cs
@@ -1,10 +1,12 @@
 ﻿
+using System;
+
 public class BankAccount
 {
     public string FullName { get; set; }
-    public int AccountNumber { get; set; }
-    public int CardNumber { get; set; }
-    public int PinCode { get; set; }
+    public Int64 AccountNumber { get; set; }
+    public Int64 CardNumber { get; set; }
+    public Int64 PinCode { get; set; }
     public decimal Balance { get; set; }
 
     public bool isLocked { get; set; }
